refactor Gateway API CRD reconciler and installation logic
